requirements.txt
python-dotenv==0.20.0 
mysql-connector-python==8.0.29   
requests==2.28.0
pip==22.1.2  
requests==2.28.0
setuptools==58.1.0
urllib3==1.26.16